Implement ceilinged/floored division to multiply transformation.
A multiplier and shift value is chosen so that ceiling(x/y) can be
converted to floor(m*x/2^s)+1, or (1+ (ASH (* M X) -S)).
This doesn't work when x = 0, so that is handled separately.
Floor(x/y) is simply transformed into floor(m*x/2^s), or (ASH (* M X) -S).
There's no need to explicitly check for 0.
In both signed floor and ceiling, the multipliers that work for positive
dividends aren't correct for negative ones and vice versa, so the
transformed expression first checks the sign of the dividend and then
uses the appropriate multiplier.

Calculate the remainder for unsigned ceiling in two steps so that inline
arithmetic is used (X - (Q-1)*Y - Y instead of (X - Q*Y)).
To do this, we need the value of the quotient before it is increased by one,
so the generator doesn't wrap the (1+ ...) around the expression (we do that
manually later in the ceiling transform). The check for X = 0 is also done
outside the generator.
For signed ceiling, calculate the remainder normally or as in the unsigned
case, depending on the sign of the result.
The floored division's remainder is calculated similarly.
The build fails for some reason when the floor transform is used in
target-char.lisp, so the floor transform gives up during compilation.