More efficient remainder calculation.
Calculate the remainder in the unsigned ceiling transform in two steps
so that inline arithmetic is used (X - (Q-1)*Y - Y instead of (X - Q*Y)).
To do this, we need the value of the quotient before it is increased by one,
so calculate the final quotient outside of GEN-UNSIGNED-DIV-BY-CONSTANT-EXPR.
Place the check for X = 0 outside too.