Add new functions: SB!KERNEL:%(SIGNED)-MULTIPLY-HIGH-AND-SHIFT.
These perform a multiply-high and a shift on signed and unsigned words.
They give a little performance boost for fixnums over calling mulhi and
shift separately.
VOPs are implemented for x86, x86-64, PPC and ARM. The PPC and ARM
code hasn't been tested yet.
Replace the multiply-high and shift calls with these functions in the
constant division transforms.