Optimize the unsigned truncate transform when the argument is a fixnum.
Instead of shifting before and after divide, incorporate the right shift
into the divisor and the left shift into the final shift value.
The lower bits get masked at the end to get the tags back.
This makes it possible to remove the TRULY-THE that was necessary to
get the fixnum MULTIPLY-HIGH VOPs.