Add new function: %LOSE-WORD-DERIVED-TYPE.
This returns its argument, but sets the derived type back to WORD. This
is needed for the fixnum optimization in GEN-UNSIGNED-DIV-BY-CONSTANT-EXPR
to remove unnecessary shifting.
Previously, before the mask was applied, the compiler converted the number
into a fixnum when its derived type allowed it, and then
%TAGGED-WORD-TO-FIXNUM converted it back, leading to two unnecessary shifts.

TODO: Find a better/cleaner way of doing this, if possible.