Extend the ceiling and floor generators for rational numbers.
Extend CHOOSE-CEILING/FLOOR-MULTIPLIER so that they be used to replace
ceilinged/floored multiplication by A/Y, and rewrite the ceiling/floor
generators to work with rational divisors too.
Instead of using (if t ...) or (if nil ...) when the dividend can only
be negative or positive, don't do any condition checking at all.
This doesn't cause any performance changes, since the comparison
would have been removed during compilation anyway, but it shortens
the code a little and makes the examples more human-readable.

Also add a new function: GEN-MULTIPLY-SHIFT-EXPR. Use calls to this
function in the div-by-mul code generators instead of duplicating code.