Add transform for unsigned truncated division by fractions.
Modify CHOOSE-MULTIPLIER so that it can be used for multiplying
by fractions, and add a code generator that creates a simpler expression
for truncated multiplication of X by the constant rational A/Y.
The multiplier can get very large, so the generator tries to perform
the operations in multiple smaller steps whenever possible. For details,
see the comments in the generator or Paul Khuong's "Fast Truncated
Multiply-Divide by Positive Constants".