Add transform for signed integer-rational truncated division.
Add new function: SB!KERNEL:%SIGNED-MULTIPLY-AND-ADD-HIGH. It multiplies
two args, adds a third arg to the product, and returns the high bits of the
sum. It is implemented with VOPs on x86(-64), ARM and PPC.
Use it to carry out the 2N-bit multiply-shift similarly as in the unsigned
case.
Also make the examples easier to read by using macros.